Trivial: rename exceptions and fee schedule
